Reorganise outfit package
