+++ src/main/java/com/jumpcutfindo/happyholidays/common/item/outfits/OutfitItem.java
@@ -1,4 +1,4 @@
-package com.jumpcutfindo.happyholidays.common.item.christmas.outfits;
+package com.jumpcutfindo.happyholidays.common.item.outfits;
 
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.item.ArmorMaterial;
