+++ src/main/java/com/jumpcutfindo/happyholidays/common/registry/christmas/ChristmasOutfits.java
@@ -1,4 +1,7 @@
-package com.jumpcutfindo.happyholidays.common.item.christmas.outfits;
+package com.jumpcutfindo.happyholidays.common.registry.christmas;
+
+import com.jumpcutfindo.happyholidays.common.item.outfits.Outfit;
+import com.jumpcutfindo.happyholidays.common.item.outfits.OutfitItem;
 
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.item.ArmorMaterial;
